refactor: replace types with interfaces
